Update README.md
I Have Added extra info to the Readme file as well as added a contribution list and a short description of the rules and what is expected in the process of contributing to this Project.

If you could please mark this as completed for Hacktober contribution that would be great! Thank you so much.